requirements.txt
dnspython==1.15.0
dnspython3==1.15.0
